+++ src/redux/actions/currencies/get-currencies-success.js
@@ -1,0 +1,6 @@
+export const GET_CURRENCIES_SUCCESS = 'GET_CURRENCIES_SUCCESS';
+
+export const getCurrenciesSuccess = payload => ({
+  type: GET_CURRENCIES_SUCCESS,
+  payload,
+});