+++ src/redux/actions/currencies/index.js
@@ -1,0 +1,3 @@
+export * from './get-currencies';
+export * from './get-currencies-failure';
+export * from './get-currencies-success';