+++ src/redux/epics/patch-usage-consolidation.js
@@ -17,7 +17,7 @@ export default ({ usageConsolidationApi }) => (action$, store) => {
 
       return usageConsolidationApi
         .patchUsageConsolidation(store.getState().okapi, credentialsId, { data })
-        .map(patchUsageConsolidationSuccess)
+        .map(() => patchUsageConsolidationSuccess(data.attributes))
         .catch(errors => Observable.of(patchUsageConsolidationFailure(errors)));
     });
 };